make CertificationPath a class and rename chain
